lock debugger: Blacklist apphelp.dll to make lock debugger work in SkyrimVR
apphelp.dll has a lock that keeps being seen as taken, but not released
by 3DMigoto's lock dependency debugger, making the output useless and
resulting in 3DMigoto hanging the game.

Blacklist this dll so that output will focus on SkyrimVR and SKSE
plugins of interest, and so that 3DMigoto will not itself cause the game
to lock up.