Full reformat of all util.cpp, util.h, util_min.h
Just the reformat from ctrl-k/d here to set the clang-format we are using.  Resharper chop for function parameter alignment on next lines, which is then respected by clang-format.

Move all pointer declartion left justified.
Braces are all added and aligned.
Function input and local vars are aligned.

A couple of added clang-format off/on to avoid reformatting struct defs that wrap to single line.  Not available until clang-format v13, and we are using v12 with VS2019.