Rename EnumName_t to Enum_Name_t
Our style guide suggest to use simply enum_name, but this struct is a little special in that it is used for a lot of Enum to Name conversions.

I didn't want to leav it as is, because our style would then suggest it's an object, and it's not.  So, a compromise is to add the _ to make it Enum_Name_t which then looks closer to our Enum_Style guide.  And since it is used specifically for Enum conversions, this seems like best compromise.

Left the trailing _t which we also normally strip, so that it is clearly not a normal Enum.

Added to clang-tidy naming as an exception to avoid red squiggles.

Full rebuild of all 6 targest is successful.