Change HuntingMode to Hunting_Mode enum.
For our style guidelines, we want enums to be Camel_Snake_Case as described by clang-tidy.  As an aside, we also want all Enums to be enum Class, scoped enums, as recommended by the C++ github.

We also want the EnumConstantCase set to lower_case, because even though these are constants, we want to avoid visually looking like MACROS.

There is an option for camel_Snake_Back in clang-tidy, but was rejected as being bad when we run into constants with acronyms like dump_rt, and dump_srv, fmt_2d_jps.

This commit changes every use of HuntingMode enum to Hunting_Mode:: because it's now a class, so we can't just use the constant itself.  So HUNTING_MODE_ENABLED -> Hunting_Mode::enabled.

Changed lots of bad uses of just (G->hunting) as a boolean, when it's actually an int.  These now need to be (G->hunting != Hunting_Mode::disabled) to get the non-zero, or mode 1,2 check.

.clang-tidy updated for EnumConstantCase.

Still compiles.