Move Hunting_Mode Enum into util.h
This is not really a global, it's an Enum definition used in a Util style.  There are quite a few of these that are logically more utility than global, so this is a first test.

Adding a new routine hunting_enabled(), as a simple utility routine to avoid the nasty negative language syntax of
(G->hunting != Hunting_Mode::disabled)
for the modes of enabled||soft_disabled.

Used in multiple places, this reads a lot better if we encapsulate that boolean in a utility routine.

Full rebuild of all 6 targets is successful.