Move Hunting_Mode def into Hunting.hpp.
These don't really make sense in Util either.  It's part of hunting, and the definitions for on/off.  They belong in the Hunting.hpp file and Hunting.cpp for the utility routine.

Move these from Util->Hunting, and update the headers for the other users, including Globals.h.

Fix the includes for Hunting.hpp to be what it actually needs, not the HackerDevice which made no sense, and set up a header loop through globals.

Compiles on all 6 targets.