Fix several clang errors so clang-tidy can run.
clang is more restrictive than the MSVC, but we need it to run without errors so that the clang-tidy can succeed.  That allows us to use automated renaming instead of having to do it manually.

This commit fixes several clang errors that use microsoft assumptions.

The routine in Override.hpp is rewritten to avoid using a switch statement, because the offsetof macro is not strictly compile time only, and thus not a constant that clang wants.