Rename Log_Level enum in Overlay.hpp.
To match our enum style guide:
 LogLevel->Log_Level
 LOG_DIRE->dire
 LOG_NOTICE->notice
 etc.

Log_Level is also now a scoped enum via class.  So in Overlay.cpp we need to use static_cast for usages of the Log_Level, since it's no longer a simple const.

This affects a lot of files, but does improve readability.  There is redundancy in the usage as:
 log_overlay(Log_Level::dire, "...
In that Log shows up twice, but we want the Enum to have an underscore.