Rebuild every .hpp header to avoid circular references.
We have long had a big header mess, and this commit fixes all of these header problems.

For every file here, I went through their actual usages, not just leftover #includes at the top, and removed everything unused, and now include only what is necessary.  For .h and .hpp files, that is strictly what the header file needs, not the corresponding cpp file.

In particular, every use of HackerDevice and HackerContext in the headers was changed to be a forward reference, so that we do not need to include those headers and set up circular references.  Since we only ever reference these objects by pointer, this works.

For the .ccp files, I went through and removed everything, then restored it piece by piece so that we have minimal and accurate includes now.

The iid.cpp file was referenced out of util.h which was not great, so I added an iid.h header for it, and moved the iid functions there.