Fully reformat Overlay.cpp.
This starts with the Resharper chop, so that function params are all on their own lines, then ctrl-k/d for the clang-format reformat, which aligns everything and left justifies * definitions and fixes brace style and whitespace.

Includes adding {} blocks around ENTER_CRITICAL so that they are reformatted as well.

All 6 build targets are successful.