Move all code out of Profiling.hpp.
Code in header files is not considered good style, so this moves everything except for the egregious template code back into the .cpp file.

In particular, all of the NvAPI_* functions are defined in the header file now, but moved into the .cpp file.

The function prototype is changed back to match the way they are defined in nvapi.h, because for any function we are wrapping or hooking, we want to maintain the exact definition.

Adds some clang-tidy override regex for the function names and nvidia vars so they won't be tagged as needing renaming or accidentally changed from a clang-tidy update.