Rename Mode enum in Profiling.hpp.
Renaming to use all lower_case for enum constants.  So:
SUMMARY->summary
NONE->none
and so on.

Renaming enum class Mode {} to mode, so that the enum class is lower case.  Usage will therefore be like:
        if (profiling::profile_type == profiling::mode::summary) {

Where the use of the namespace is the key part that specifies it is an enum usage.  This keeps it from looking like object references which we want to reserve for actual objects.