Renaming namespace Profiling to profiling.
We want namespaces to be lower_case so that they do not visually conflict with Classes.  This changes all uses of Profiling to lower case profiling, so the usage is now looking like:
        if (profiling::profile_type == profiling::mode::summary) {
which makes it easier to distinguish that this is namespace+enum usage of the scope operator and not a class level usage.