Move Log_Level to namespace overlay::log in Overlay.hpp.
Our enums look better and are more clear if we put them into a namespace.  This is also considered best practices, to avoid having stuff in the global namespace.

Usage changes from:
 log_overlay(Log_Level::dire, "BUG: Invalid...
to
 using namespace overlay;
 log_overlay(log::dire, "BUG: Invalid...

Where we are using at the top of each file. It's also OK to just use the full scope overlay::log::dire, but that is a little verbose for these obvious logging lines.  But is left in for the rare uses of num_levels, where it's not obvious it's part of overlay.

We also move the functions of log_overlay_w, log_overlay, clear_notices into the namespace overlay.  So they will normally be called as overlay::logoverlay() to make it more clear what functionality is used.

This is part of refactoring the enum usage to make it more clear at use time.

All 6 build targets are still successful.