Move ini_parser_lite to new ini_parser_lite.h
Move these out of min_util.h, because we don't currently have an ini_parser_lite.h, and it makes more sense to keep those distinct.

Util is too much of a dumping ground, and the min_util is a good example of it not working and needing a new set.

The minimal ini_parser_lite is not unreasonable, and now is it's own unit.