Update all .h and .hpp files using Windows.h as root.
Since we are stuck on Windows, using the Windows.h is the stackoverflow recommended approach, instead of all the ancillary files it includes.  We previously accidentally found stuff like the LONGPTR, IID, or HWND definitions, because they were included by other files.

This pass through the headers should finalize all these .h and .hpp files as the exact set of headers necessary for that unit.  No circular references.

This should now be a robust set of header files that won't accidentally break if a different file changes its includes.

Every header file was reviewed:
#include "Assembler.h"
#include "Cursor.h"
#include "DecompileHLSL.h"
#include "DrawCallInfo.h"
#include "EnumNames.hpp"
#include "HackerDXGI.hpp"
#include "HookedContext.h"
#include "HookedDXGI.h"
#include "HookedDevice.h"
#include "iid.h"
#include "Lock.h"
#include "log.h"
#include "Hunting.hpp"
#include "IniHandler.h"
#include "ini_parser_lite.h"
#include "Input.hpp"
#include "NVProfile.h"
#include "Overlay.hpp"
#include "shader.h"
#include "util.h"
#include "vkeys.h"
#include "D3D11Wrapper.h"
#include "DLLMainHook.h"
#include "HackerContext.hpp"
#include "Profiling.hpp"
#include "ResourceHash.hpp"
#include "ShaderRegex.hpp"
#include "Override.hpp"
#include "HackerDevice.hpp"
#include "Globals.h"
#include "FrameAnalysis.hpp"
#include "CommandList.hpp"