Finalize all #include in .cpp files.
This is an extensive rebuild of all the header files to bring them up to best practices.  Lots of files had magic includes, which tends to be a fragile mechanism for any changes, and now every file has a list of all directly used includes.  This also serves partly as documentation, showing what each file actually is using.

Basic starting point was to use <Windows.h> as the recommended include for windows dev.  It includes far more than we need, but picking and choosing windows headers was discouraged.

All #includes are sorted and structured around having most global includes last, as recommended on StackOverflow. This is automatic, as part of clang-format.

All includes were moved to the top, having stuff midstream is confusing and misleading.

The use of system Guids was resolved by avoiding initguid.h and its first in line requirement, and linking directly with dxguid.lib.

"nvapi.h" was moved after <Windows.h> to be last, so that the a prior <d3d11_1.h> header can define the __d3d11_h__ macro.  All <d3d11.h> were moved to <d3d11_1.h>

This is tedious and error prone to work out, so it's possible there are still missed includes here, but this should be pretty good.

All 6 build targets are successful.