Disable functionality when [Loader]Target doesn't match exe name
The [Loader]Target setting already exists used when the game is loaded
via 3DMigoto Loader.exe from outside of the game directory, and in that
case 3DMigoto will fail DllMain so that it won't be loaded into
unintended targets. However when geo-11 is in the game directory we
can't do that, because that would cause the dynamic linker to fail
loading the game, and so 3DMigoto/geo-11 will always load in that case
even if the target doesn't match.

For the most part that wasn't a big deal in 3DMigoto with 3D Vision as
the output as basically nothing would look wrong, but in geo-11 some of
the output modes are problematic if they run in a windowed launcher or
similar.

This commit adds a second check later in InitializeDll() to make sure
[Loader]Target matches the exe name, and instead of failing DllMain and
unloading the DLL, it will leave the DLL loaded, but pass through calls
to the real DirectX without the usual geo-11 processing.

Backporting this to 3DMigoto anyway, I recall some people running into
issues with 3DMigoto causing issues in launchers that this could help
with.