Add effective_dpi for UI scaling
This will return the "effective DPI" (not the real/raw/physical DPI)
that Windows reports for the purpose of scaling UI elements (such as the
help text shader) on ultra high resolution / 4k displays.

It will attempt to retrieve the effective DPI for the monitor that the
game is displayed on, though the Windows APIs for this have taken
multiple iterations to get right, so this behaviour might end up being
somewhat dependent on a combination of the game's process wide DPI
awareness, the version of Windows it is running on and the resolution...
Wider testing may be needed to ensure that this behaves well in as many
situations as possible.

The APIs are loaded dynamically to ensure Windows 7 compatibility
remains intact, and on these older versions that don't support effective
DPI it will return 96, which is the effective DPI reported when scaling
is set to 100% or to DPI unaware applications.