HLSLDecompiler: Fix use of overwritten index variable
Indexing into matrices works differently in assembly and HLSL - in
assembly the index is in multiples of float4s, but in HLSL it is in
multiples of matrices.

The decompiler has a few strategies to handle this difference, notably
it attempts to recognise where the index variable in assembly was
multiplied by the number of rows (columns?) in the matrix, and tries to
use the input register of that multiplication as the index instead of
the output register.

However, in the event that the input and output to that multiplication
were the same register it will use the output variable that has been
multiplied in assembly, and will then be implicitly multiplied again in
HLSL.

This adds a check to skip this strategy if the input and output
registers to the multiplication are the same. The decompiler will then
select the next strategy, which is dividing the index variable by the
number of rows in the matrix.