Add draw=auto
This works like the existing drawindexed=auto and drawindexedinstanced=
auto commands, but calculates the vertex count from the size and stride
of the currently bound vb0.

This is useful for Unity games where the pre-skinning shaders run on the
mesh vertex buffers without an index buffer assigned where drawindexed=
auto would not be suitable.

Some relevant discussion happening over here:
https://github.com/DarkStarSword/3d-fixes/issues/19