Add D3D11On12CreateDevice to exports.
Add a passthrough function for any calls to D3D11On12CreateDevice. This is a new static export for a d3d11.dll, and without it we can get games that will refuse to run because of the missing entry point.

This is not the same as the DXGI hook for the function, and I'm not sure how this will interact with that hook.