Add a zip_release.yml for CI builds.
(copied from vs0222 branch- this is necessary in master for Actions to be available. It is currently set for only manual builds.)

This is an edited copy of the geo-11 CI build, which uses the Publish.bat but properly auto-increments the build version number.

This should setup a vs2022 environment and do a full zip release build, leaving artifacts on github for downloads.

We are done making builds on our own machines, the actions are far superior.