Checkpoint as a force push.
Did a (generally bad) force push here, to rewrite the *public* history of the branch. No one will be using this branch for active development, so it's OKish.

I did this to remove a couple of merges from master into this branch, which breaks up the history for the edit/cleanup/renames.

Not super important, but after I finalize the renaming in a week or so, I will merge back master, and pick up all the 10 or so remaining non-merged items too.