Change all 'using namespace std' to specific 'using std::vector' types.
Avoiding 'using namespace std' altogether, in cpp files.

This changes the headings of the cpp files to use the common types:
  std::vector;
  std::string;
  std::wstring;
and so on.

Rare types like std::exception, or std::wistringstream, are still called out explicitly in the code.

Removes all uses of common inline stuff like std::string, std::vector.

In .h files, we leave them as explicit, as it's more important to know exactly what the type is, and where it comes from. We want to not use any using clause in .h or .hpp files, as they get folded into cpp compiles.

Code compiles but is untested.