Bump SDK version to 10.0.19041.0
Time to update the tool sets.  Let's start with the Windows SDK as being the least likely to cause any problems.

Lots of options, but ChatGPT strongly suggests that 19041 is the most compatible and best tested version that still supports Win7.

We care a tiny bit about Win7 support, because we might need to run in the Win7 compatibility layer. And we don't need or want anything from Win11.

This compiles all projects. But did not test running.