Bump VS version to 17 for VS2022.
Setting .sln header to Version 17, so that by default we will open in VS2022. Also bump all our .vcxproj files to v17 for consistency.

It's hard to find a download copy of VS2019 now, and there is no particular drawback to using VS2022. There are some good Git enhancements that are worthwhile.

Builds and tested as working in Ixion.