Add DirectXTK v2025.7.10 as static library.
We'll just build against the SDK style static library for this piece now. Files were fetched from the vcpkg static distribution, not built locally.

Lots of pros and cons to weigh, but current best approach seems to be to just use SDK style libraries, that are directly added to the project. These will all reside in ThirdPartyLibs folder.