Fix build to use new ThirdPartyLibs folder.
Files come from the static vcpkg distribution, not locally built. We are using the static distribution to match our use of /MT through all other parts of project.

Moved in a function from the source code for GetSurfaceInfo, as opposed to doing a code-drop for latest version.

All 6 build targets successfully build now, but are untested.