Revert "Some small tweaks to compile under c++17."
This reverts commit 19e636a78e6ee2a08b8f19bb65a7f01187380363.

Turns out these changes for C++17 are not correct, and thus we don't want to switch to C++17 just yet. Because Rename_Reformat drastically changes the headers, we'll revisit moving to C++20 after that is finished.

One key point is that the languages and toolsets have interoperability, so it's not that important to have them all match.