Remove obsolete and missing link dependencies.
These were set up to AdditionalDependency for:
$(OutDir)nvapi.lib
which is really weird, and not a good idea.
I'm really not sure where this came in, but now that the output directory has changes, this no longer links.

Removing all 6 targets.  Batch build of all 6 targets is successful, but dll is untested.